+++ pkg/webui/console/views/device-payload-formatters/index.js
@@ -24,12 +24,15 @@ import Tab from '@ttn-lw/components/tabs'
 import Notification from '@ttn-lw/components/notification'
 
 import NotFoundRoute from '@ttn-lw/lib/components/not-found-route'
+import withRequest from '@ttn-lw/lib/components/with-request'
 
 import DeviceUplinkPayloadFormatters from '@console/containers/device-payload-formatters/uplink'
 import DeviceDownlinkPayloadFormatters from '@console/containers/device-payload-formatters/downlink'
 
 import sharedMessages from '@ttn-lw/lib/shared-messages'
 import PropTypes from '@ttn-lw/lib/prop-types'
+
+import { getApplicationLink } from '@console/store/actions/link'
 
 import {
   selectApplicationLink,
@@ -46,16 +49,21 @@ const m = defineMessages({
   infoDownlinkText:
     'These payload formatters are executed on downlink messages to this end device and take precedence over application level payload formatters.',
 })
-@connect(state => {
-  const link = selectApplicationLink(state)
-  const fetching = selectApplicationLinkFetching(state)
+@connect(
+  state => {
+    const link = selectApplicationLink(state)
+    const fetching = selectApplicationLinkFetching(state)
 
-  return {
-    appId: selectSelectedApplicationId(state),
-    devId: selectSelectedDeviceId(state),
-    fetching: fetching || !link,
-  }
-})
+    return {
+      appId: selectSelectedApplicationId(state),
+      devId: selectSelectedDeviceId(state),
+      fetching: fetching || !link,
+    }
+  },
+  {
+    getApplicationLink,
+  },
+)
 @withBreadcrumb('device.single.payload-formatters', props => {
   const { appId, devId } = props
   return (
@@ -65,6 +73,7 @@ const m = defineMessages({
     />
   )
 })
+@withRequest(({ getApplicationLink, appId }) => getApplicationLink(appId, ['default_formatters']))
 export default class DevicePayloadFormatters extends Component {
   static propTypes = {
     location: PropTypes.location.isRequired,